lawyer/LinhCSE/DistilCSE_sym/requirements.txt
clip==0.2.0
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.2
pandas==1.3.3
prefetch_generator==1.0.1
prettytable==2.2.1
scikit_learn==0.24.2
scipy==1.5.4
setuptools==49.3.0
skipthoughts==0.0.1
torch==1.10.0
tqdm==4.49.0
transformers==4.11.3
info-nce-pytorch==0.1.4
focal-loss-torch==0.1.2
scikit-learn==0.24.2
sentence-transformers==2.2.2
# pip install torch==1.10.0+cu111 torchvision==0.11.0+cu111 torchaudio==0.10.0 -f https://download.pytorch.org/whl/torch_stable.html
